@@ -8,27 +8,29 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "^14.0.0",
+        "@angular/animations": "^15.2.9",
         "@angular/cdk": "^15.2.3",
-        "@angular/common": "^14.0.0",
-        "@angular/compiler": "^14.0.0",
-        "@angular/core": "^14.0.0",
-        "@angular/forms": "^14.0.0",
+        "@angular/common": "^15.2.9",
+        "@angular/compiler": "^15.2.9",
+        "@angular/core": "^15.2.9",
+        "@angular/forms": "^15.2.9",
         "@angular/material": "^15.2.3",
-        "@angular/platform-browser": "^14.0.0",
-        "@angular/platform-browser-dynamic": "^14.0.0",
-        "@angular/router": "^14.0.0",
+        "@angular/platform-browser": "^15.2.9",
+        "@angular/platform-browser-dynamic": "^15.2.9",
+        "@angular/router": "^15.2.9",
         "@ngneat/reactive-forms": "4.0.4",
         "@rxweb/reactive-form-validators": "2.1.6",
         "jssha": "^3.3.0",
+        "lodash": "^4.17.21",
         "rxjs": "~7.5.0",
         "tslib": "^2.3.0",
         "zone.js": "~0.11.4"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^14.0.1",
-        "@angular/cli": "~14.0.1",
-        "@angular/compiler-cli": "^14.0.0",
-        "typescript": "~4.7.2"
+        "@angular-devkit/build-angular": "^15.2.8",
+        "@angular/cli": "~15.2.8",
+        "@angular/compiler-cli": "^15.2.9",
+        "@types/lodash": "^4.14.194",
+        "typescript": "~4.9.5"
     }
 }
