@@ -4,8 +4,10 @@ import {BrowserModule} from '@angular/platform-browser';
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
 import {NavbarComponent} from "./presentation/components/navbar/navbar.component";
-import {HttpClientModule} from "@angular/common/http";
-import {fakeBackend} from "./shared/providers/fake-backend";
+import {HTTP_INTERCEPTORS, HttpClientModule} from "@angular/common/http";
+import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
+import {MatSnackBar} from "@angular/material/snack-bar";
+import {FakeBackendService} from "./db/fake-backend.service";
 
 export const STORAGE = new InjectionToken<Storage>('Local storage token');
 
@@ -18,11 +20,12 @@ export const STORAGE = new InjectionToken<Storage>('Local storage token');
         AppRoutingModule,
         NavbarComponent,
         HttpClientModule,
-        // HttpClientInMemoryWebApiModule.forRoot(DbService)
+        BrowserAnimationsModule,
     ],
     providers: [
         {provide: STORAGE, useValue: localStorage},
-        fakeBackend,
+        {provide: HTTP_INTERCEPTORS, useClass: FakeBackendService, multi: true},
+        MatSnackBar,
     ],
     bootstrap: [AppComponent]
 })
