@@ -1,7 +1,7 @@
 import {inject, Injectable} from '@angular/core';
 import {Observable, of} from "rxjs";
 import {ProfileService} from "./shared/services/profile.service";
-import {PROFILE} from "./shared/consts";
+import {PROFILE_KEY} from "./shared/consts";
 import {STORAGE} from "./app.module";
 import {AuthService} from "./shared/services/auth.service";
 
@@ -19,7 +19,7 @@ export class AppInitializerService {
     ) {}
 
     init(): Observable<boolean> {
-        const storageUser = this.storage.getItem(PROFILE);
+        const storageUser = this.storage.getItem(PROFILE_KEY);
         if (storageUser) {
             this.profile.data = JSON.parse(storageUser);
             this.auth.isLogin = true;
