@@ -1,18 +1,19 @@
 import {Component} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {AuthService} from "../../../shared/services/auth.service";
-import {RouterLinkActive, RouterLinkWithHref} from "@angular/router";
+import {RouterLinkActive, RouterLink} from "@angular/router";
 import {ProfileService} from "../../../shared/services/profile.service";
 import {User} from "../../../core/user/user.model";
 import {NavigatorService} from "../../../shared/services/navigator.service";
 import {takeUntil} from "rxjs";
 import {DestroyerService} from "../../../shared/services/destroyer.service";
 import {MatButtonModule} from "@angular/material/button";
+import {MatIconModule} from "@angular/material/icon";
 
 @Component({
     selector: 'app-navbar',
     standalone: true,
-    imports: [CommonModule, RouterLinkWithHref, RouterLinkActive, MatButtonModule],
+    imports: [CommonModule, RouterLink, RouterLinkActive, MatButtonModule, MatIconModule],
     templateUrl: './navbar.component.html',
     styleUrls: ['./navbar.component.less'],
     providers: [DestroyerService]
@@ -31,7 +32,7 @@ export class NavbarComponent {
             .subscribe(u => this.user = u);
     }
 
-    logout(): void {
+    protected logout(): void {
         this.auth.logout();
         this.navigatorService.toLoginPage();
     }
