@@ -4,6 +4,11 @@ import {CommonModule} from '@angular/common';
 import {LoginRoutingModule} from './login-routing.module';
 import {LoginComponent} from './login.component';
 import {ReactiveFormsModule} from "@angular/forms";
+import {MatFormFieldModule} from "@angular/material/form-field";
+import {MatInputModule} from "@angular/material/input";
+import {MatButtonModule} from "@angular/material/button";
+import {MatIconModule} from "@angular/material/icon";
+import {MatProgressSpinnerModule} from "@angular/material/progress-spinner";
 
 
 @NgModule({
@@ -13,7 +18,12 @@ import {ReactiveFormsModule} from "@angular/forms";
     imports: [
         CommonModule,
         LoginRoutingModule,
-        ReactiveFormsModule
+        ReactiveFormsModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatButtonModule,
+        MatIconModule,
+        MatProgressSpinnerModule
     ]
 })
 export class LoginModule {
