@@ -1,6 +1,5 @@
-export const PROFILE = 'profile';
+export const PROFILE_KEY = 'profile';
 export const USERS_KEY = 'test-case-users'
-export const IS_LOGIN = 'isLogin';
 
 export const API_USERS = 'users';
 export const API_LOGIN = 'login';
